requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
assemblyai>=0.25.0
google-generativeai==0.3.2
jinja2==3.1.2
websockets==12.0
tavily-python==0.3.3
